api/requirements.txt
django==4.2.4
django_extensions==3.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-auth-ldap==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-csv==2.1.1
drf-schema-adapter==3.0.6
colorful-logger==0.1.13
gunicorn==21.2.0
uvicorn==0.23.0
psycopg2==2.9.6
pyyaml==6.0.1
pandas==2.0.3
numpy==1.25.2
tqdm==4.65.0
easydict==1.10
schema==0.7.5
requests==2.31.0
shortuuid==1.0.11



